[G-11] 서버 구현하지 않는 관계로 폴더 구조 변경 (#21)
- [refactor] 폴더 구조 변경 - client 제거 및 1 depth 상위에서 관리
- [refactor] src 폴더 구조 변경 - apis, styles 폴더 구조 변경
