+++ src/components/PostItem/styles.ts
