+++ src/components/Posts/styles.ts
