+++ src/pages/_app.tsx
@@ -1,4 +1,3 @@
-// import '@/styles/global.css';
 import type { AppProps } from 'next/app';
 
 import { Global } from '@emotion/react';
